+++ mmseg/utils/__init__.py
@@ -13,9 +13,9 @@ from .collect_env import collect_env
 from .io import datafrombytes
 from .misc import add_prefix, stack_batch
 from .set_env import register_all_modules
-from .typing import (ConfigType, ForwardResults, MultiConfig, OptConfigType,
-                     OptMultiConfig, OptSampleList, SampleList, TensorDict,
-                     TensorList)
+from .typing_utils import (ConfigType, ForwardResults, MultiConfig,
+                           OptConfigType, OptMultiConfig, OptSampleList,
+                           SampleList, TensorDict, TensorList)
 
 __all__ = [
     'collect_env', 'register_all_modules', 'stack_batch', 'add_prefix',
